refactor(level): move level package from kernel to core/logger subpackage
level declares Debug/Info/Warn/Error — that is logger-domain
vocabulary, not a generic primitive. The kernel rule is 'stdlib-only
AND generic'; level only satisfied the first half. No future
non-logger domain will ever import level.

Audit of the other kernel packages (errs, clock, buffer) shows they
all remain defensible in kernel: errs is meta-infrastructure used by
every layer, clock is a textbook generic time abstraction, buffer is
a generic byte-pool scratchpad that any future byte-oriented domain
could reach for.

Mechanical move:
- internal/kernel/level/        -> internal/core/logger/level/
- 5 production imports + 4 test imports flip their path literals
- ADR 0002 registry row 1100-1199 owner updated
- Package-level READMEs touched to reflect the new layer
- depguard rules unchanged: level moved DOWN the allowed chain
  (kernel -> core), core-deps / service-deps / pkg-deps already
  permit core imports

Public API unchanged: pkg/v1/logger.Level and LevelDebug/Info/Warn/
Error keep their identity via type aliases; only the underlying
import path in the SDK's internal source differs.

Tests: all race-clean with coverage matching the previous run (90-
100% on emitter modules, 100% on the relocated level package).
ktn-linter clean.
