+++ internal/service/logger/logger_external_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 
 	corelogger "github.com/kitsunium/sdk/internal/core/logger"
+	"github.com/kitsunium/sdk/internal/core/logger/level"
 	"github.com/kitsunium/sdk/internal/kernel/errs"
-	"github.com/kitsunium/sdk/internal/kernel/level"
 	svclogger "github.com/kitsunium/sdk/internal/service/logger"
 )
