+++ internal/service/logger/logger_internal_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	corelogger "github.com/kitsunium/sdk/internal/core/logger"
-	"github.com/kitsunium/sdk/internal/kernel/level"
+	"github.com/kitsunium/sdk/internal/core/logger/level"
 )
 
 // mustNewTextHandler builds a TextHandler or fails the test. Shared helper
