+++ pkg/v1/logger/logger.go
@@ -11,7 +11,7 @@ import (
 	"os"
 
 	corelogger "github.com/kitsunium/sdk/internal/core/logger"
-	"github.com/kitsunium/sdk/internal/kernel/level"
+	"github.com/kitsunium/sdk/internal/core/logger/level"
 	svclogger "github.com/kitsunium/sdk/internal/service/logger"
 )
